backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
beanie>=1.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.36.0
requests>=2.31.0
python-dotenv>=1.0.1
